+++ CHANGELOG.md
@@ -30,6 +30,7 @@ in the detailed section referring to by linking pull requests or issues.
 * Add contract validation rule (#1239)
 * Harmonize setting names in `data-plane-transfer` (#1164)
 * Add component tests coverage to the codecov coverage report (#1246)
+* Postgresql end to end test (#1278)
 
 #### Changed
 
@@ -41,6 +42,7 @@ in the detailed section referring to by linking pull requests or issues.
 * Make data-plane public api controller asynchronous (#1228)
 * Provide In-mem implementations by default (#1130)
 * Changed Catalog config keys and switched from minutes to seconds
+* Uniform all the sql schema files as `docs/schema.sql` (#1278)
 
 #### Removed